+++ app/javascript/components/If.jsx
@@ -1,0 +1,3 @@
+export default function If(props) {
+    return !!props.condition && props.children;
+}
