+++ app/javascript/components/RedditPreview.jsx
@@ -1,0 +1,7 @@
+import { h, render, Fragment } from "preact";
+
+export default function RedditPreview(props) {
+    const data = props.data;
+
+    return (<p>Coming soon</p>);
+}
